Interactables

+++ Assets/Scripts/Items/Chest.cs
@@ -1,10 +1,17 @@
 using System;
 using UnityEngine;
 
-public class Chest : HidingPlace, IHideable
+public class Chest : HidingPlace
 {
+    private GameObject _lid;
+
     private Interactable _curHandItem;
 
+    protected override void Start()
+    {
+        base.Start();
+        _lid = transform.GetChild(1).gameObject;
+    }
     /**
      * if chest is open:
      *     if we have wrench as a hand object- we can close the chest 
@@ -14,29 +21,30 @@ public class Chest : HidingPlace, IHideable
      */
     public override Action[] CalcInteractions()
     {
-        _curHandItem = GameManager.Instance.inventory.GetHandItem();
+        // Handling the main inventory 
+        MainInventory main = GameManager.Instance.inventory;
+
+        _curHandItem = main.GetHandItem();
         
-        // chest is open
-        if (open)
+        // we don't carry a non-hidable object
+        if (!(_curHandItem != null && !(_curHandItem is IHideable)))
         {
+            
             // we have a wrench in the hand
-            if (GameManager.Instance.inventory.IsInInventory(ItemType.Wrench, true))
+            if (GameManager.Instance.inventory.IsInInventory(ItemType.Wrench))
             {
-                return new Action[] {Close, Hide, Exchange};
+                if (open)
+                    return new Action[] {PickUp, Close, Exchange};
+                return new Action[] { PickUp, Open };
             }
             
-            // we have another item in hand
-            if (_curHandItem != null)
-            {
-                return new Action[] {Hide, Exchange};
-            }
-            // Chest is open and there is nothing to hide
-            return new Action[] {Exchange};
+            if (open)
+                return new Action[] { PickUp, Exchange };
         }
 
-        // chest is closed
+        // either player already carries a non-hidable item or the chest is closed
         // don't need to check if there's a place for the items because this is a hand items and can always be picked up
-        return new Action[] {PickUp, Open};
+        return new Action[] {PickUp};
     }
     
     /**
@@ -46,16 +54,7 @@ public class Chest : HidingPlace, IHideable
     {
         Debug.Log("Pickup");
         
-        Interactable previousHandItem = GameManager.Instance.inventory.AddItem(this, true);
-
-        // had a hand object already- place it where the item we picked up was 
-        if (previousHandItem != null)
-        {
-            GameObject previousHandItemGM = previousHandItem.gameObject; 
-            previousHandItemGM.SetActive(true);
-            previousHandItemGM.transform.position = transform.position;
-        }
-        
+        GameManager.Instance.inventory.AddItem(this);
     }
     
     /**
@@ -63,8 +62,9 @@ public class Chest : HidingPlace, IHideable
      */
     public override void Open() 
     {
+        spriteIndex = 0;
         Debug.Log("Open chest");
-        
+        _lid.SetActive(true);
         open = true;
     }
     
@@ -73,8 +73,9 @@ public class Chest : HidingPlace, IHideable
      */
     public override void Close()
     {
+        spriteIndex = 1;
         Debug.Log("Close chest");
-        
+        _lid.SetActive(false);
         open = false;
     }
 
@@ -85,8 +86,11 @@ public class Chest : HidingPlace, IHideable
     {
         Debug.Log("Hide hand item in chest");
 
-        GameManager.Instance.inventory.DeleteItem(_curHandItem.GetItemType()); // remove from global inventory
-        hpInventory.AddItem(_curHandItem, false); // add to local inventory
+        if (hpInventory.CanAdd())
+        {
+            GameManager.Instance.inventory.DeleteItem(_curHandItem.GetItemType()); // remove from global inventory
+            hpInventory.AddItem(_curHandItem); // add to local inventory
+        }
     }
     
     /**
@@ -95,11 +99,12 @@ public class Chest : HidingPlace, IHideable
     public override void Exchange()
     {
         Debug.Log("Show the items in the chest");
-        
+
         // Handling the local inventory 
         // todo - the problem is in the load function 
-        hpInventory._inventoryUI.Load_Inventory("Chest", hpInventory); // load the items to the inventory 
-        
+        //hpInventory._inventoryUI.Load_Inventory("Chest", hpInventory); // load the items to the inventory 
+        hpInventory.StartExchange("Chest");
+
         // Handling the main inventory 
         MainInventory main = GameManager.Instance.inventory;
